Fix navigation using small screen menu:
* Move shared layout (of '/inventory' entity lists) into an 'InventoryManagement'
* Use aforementioned component instead of the shared layout (fixing the navigation blocking)
